+++ libraries/functions.rb
@@ -1,5 +1,5 @@
-class Chef
-  class Recipe
+module Cookbook #Use Cookbook name
+  module Functions #Module name
     def my_function()
       # ... your code here ...
     end