+++ recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: cookbook-example
+# Cookbook Name:: example
 # Recipe:: default
 #
 # Copyright 2016, YOUR_COMPANY_NAME